+++ packages/Webkul/Noon/src/Routes/noon-routes.php
@@ -14,28 +14,28 @@ Route::group(['middleware' => ['web', 'admin', 'tenant'], 'prefix' => config('ap
     Route::prefix('noon')->group(function () {
 
         Route::controller(CredentialController::class)->prefix('credentials')->group(function () {
-            Route::get('', 'index')->name('noon.credentials.index');
-            Route::post('create', 'store')->name('noon.credentials.store');
-            Route::get('edit/{id}', 'edit')->name('noon.credentials.edit');
-            Route::put('update/{id}', 'update')->name('noon.credentials.update');
-            Route::delete('delete/{id}', 'destroy')->name('noon.credentials.delete');
+            Route::get('', 'index')->name('admin.noon.credentials.index');
+            Route::post('create', 'store')->name('admin.noon.credentials.store');
+            Route::get('edit/{id}', 'edit')->name('admin.noon.credentials.edit');
+            Route::put('update/{id}', 'update')->name('admin.noon.credentials.update');
+            Route::delete('delete/{id}', 'destroy')->name('admin.noon.credentials.delete');
         });
 
         Route::prefix('export')->group(function () {
             Route::controller(SettingController::class)->prefix('settings')->group(function () {
                 Route::get('{id}', 'index')->name('admin.noon.settings');
-                Route::post('create', 'store')->name('noon.export-settings.create');
+                Route::post('create', 'store')->name('admin.noon.export-settings.create');
             });
             Route::controller(MappingController::class)->prefix('mapping')->group(function () {
                 Route::get('{id}', 'index')->name('admin.noon.export-mappings');
-                Route::post('create', 'store')->name('noon.export-mappings.create');
+                Route::post('create', 'store')->name('admin.noon.export-mappings.create');
             });
         });
 
         Route::prefix('import')->group(function () {
             Route::controller(ImportMappingController::class)->prefix('mapping')->group(function () {
                 Route::get('{id}', 'index')->name('admin.noon.import-mappings');
-                Route::post('create', 'store')->name('noon.import-mappings.create');
+                Route::post('create', 'store')->name('admin.noon.import-mappings.create');
             });
         });
 
@@ -43,11 +43,11 @@ Route::group(['middleware' => ['web', 'admin', 'tenant'], 'prefix' => config('ap
             Route::get('get-attribute', 'listAttributes')->name('admin.noon.get-attribute');
             Route::get('get-image-attribute', 'listImageAttributes')->name('admin.noon.get-image-attribute');
             Route::get('get-gallery-attribute', 'listGalleryAttributes')->name('admin.noon.get-gallery-attribute');
-            Route::get('get-noon-credentials', 'listNoonCredential')->name('noon.credential.fetch-all');
-            Route::get('get-noon-channel', 'listChannel')->name('noon.channel.fetch-all');
-            Route::get('get-noon-currency', 'listCurrency')->name('noon.currency.fetch-all');
-            Route::get('get-noon-locale', 'listLocale')->name('noon.locale.fetch-all');
-            Route::get('get-noon-attrGroup', 'listAttributeGroup')->name('noon.attribute-group.fetch-all');
+            Route::get('get-noon-credentials', 'listNoonCredential')->name('admin.noon.credential.fetch-all');
+            Route::get('get-noon-channel', 'listChannel')->name('admin.noon.channel.fetch-all');
+            Route::get('get-noon-currency', 'listCurrency')->name('admin.noon.currency.fetch-all');
+            Route::get('get-noon-locale', 'listLocale')->name('admin.noon.locale.fetch-all');
+            Route::get('get-noon-attrGroup', 'listAttributeGroup')->name('admin.noon.attribute-group.fetch-all');
             Route::get('get-noon-family', 'listNoonFamily')->name('admin.noon.get-all-family-variants');
         });
