+++ packages/Webkul/WooCommerce/src/Routes/woocommerce-routes.php
@@ -14,28 +14,28 @@ Route::group(['middleware' => ['web', 'admin', 'tenant'], 'prefix' => config('ap
     Route::prefix('woocommerce')->group(function () {
 
         Route::controller(CredentialController::class)->prefix('credentials')->group(function () {
-            Route::get('', 'index')->name('woocommerce.credentials.index');
-            Route::post('create', 'store')->name('woocommerce.credentials.store');
-            Route::get('edit/{id}', 'edit')->name('woocommerce.credentials.edit');
-            Route::put('update/{id}', 'update')->name('woocommerce.credentials.update');
-            Route::delete('delete/{id}', 'destroy')->name('woocommerce.credentials.delete');
+            Route::get('', 'index')->name('admin.woocommerce.credentials.index');
+            Route::post('create', 'store')->name('admin.woocommerce.credentials.store');
+            Route::get('edit/{id}', 'edit')->name('admin.woocommerce.credentials.edit');
+            Route::put('update/{id}', 'update')->name('admin.woocommerce.credentials.update');
+            Route::delete('delete/{id}', 'destroy')->name('admin.woocommerce.credentials.delete');
         });
 
         Route::prefix('export')->group(function () {
             Route::controller(SettingController::class)->prefix('settings')->group(function () {
                 Route::get('{id}', 'index')->name('admin.woocommerce.settings');
-                Route::post('create', 'store')->name('woocommerce.export-settings.create');
+                Route::post('create', 'store')->name('admin.woocommerce.export-settings.create');
             });
             Route::controller(MappingController::class)->prefix('mapping')->group(function () {
                 Route::get('{id}', 'index')->name('admin.woocommerce.export-mappings');
-                Route::post('create', 'store')->name('woocommerce.export-mappings.create');
+                Route::post('create', 'store')->name('admin.woocommerce.export-mappings.create');
             });
         });
 
         Route::prefix('import')->group(function () {
             Route::controller(ImportMappingController::class)->prefix('mapping')->group(function () {
                 Route::get('{id}', 'index')->name('admin.woocommerce.import-mappings');
-                Route::post('create', 'store')->name('woocommerce.import-mappings.create');
+                Route::post('create', 'store')->name('admin.woocommerce.import-mappings.create');
             });
         });
 
@@ -43,11 +43,11 @@ Route::group(['middleware' => ['web', 'admin', 'tenant'], 'prefix' => config('ap
             Route::get('get-attribute', 'listAttributes')->name('admin.woocommerce.get-attribute');
             Route::get('get-image-attribute', 'listImageAttributes')->name('admin.woocommerce.get-image-attribute');
             Route::get('get-gallery-attribute', 'listGalleryAttributes')->name('admin.woocommerce.get-gallery-attribute');
-            Route::get('get-woocommerce-credentials', 'listWooCommerceCredential')->name('woocommerce.credential.fetch-all');
-            Route::get('get-woocommerce-channel', 'listChannel')->name('woocommerce.channel.fetch-all');
-            Route::get('get-woocommerce-currency', 'listCurrency')->name('woocommerce.currency.fetch-all');
-            Route::get('get-woocommerce-locale', 'listLocale')->name('woocommerce.locale.fetch-all');
-            Route::get('get-woocommerce-attrGroup', 'listAttributeGroup')->name('woocommerce.attribute-group.fetch-all');
+            Route::get('get-woocommerce-credentials', 'listWooCommerceCredential')->name('admin.woocommerce.credential.fetch-all');
+            Route::get('get-woocommerce-channel', 'listChannel')->name('admin.woocommerce.channel.fetch-all');
+            Route::get('get-woocommerce-currency', 'listCurrency')->name('admin.woocommerce.currency.fetch-all');
+            Route::get('get-woocommerce-locale', 'listLocale')->name('admin.woocommerce.locale.fetch-all');
+            Route::get('get-woocommerce-attrGroup', 'listAttributeGroup')->name('admin.woocommerce.attribute-group.fetch-all');
             Route::get('get-woocommerce-family', 'listWooCommerceFamily')->name('admin.woocommerce.get-all-family-variants');
         });
